requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.2.0
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
huggingface-hub==0.20.2
hydra-core==1.3.2
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.3.2
kaggle==1.6.2
lightning==2.1.3
lightning-utilities==0.10.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
omegaconf==2.3.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pluggy==1.3.0
protobuf==4.25.2
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytest==7.4.4
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.16.2
webencodings==0.5.1
yarl==1.9.4

# Data Storage 
dvc >= 3.38.1
dvc[gdrive]
dvc[gs]




